changed value_mapper() in row 189

+++ mppsteel/model/tco_abatement_switch.py
@@ -187,10 +187,10 @@ def map_region_tco_to_plants(
     combined_df = steel_plant_ref.join(opex_capex_ref_c, how="left").reset_index()
     # Add rule to say if switch tech is a low carbon tax, apply the green premium, else apply zero
     def value_mapper(row):
-        opex = row["capex_value"] + row["discounted_opex"]
+        opex = float(row["capex_value"] + row["discounted_opex"])
         if row.switch_tech in LOW_CARBON_TECHS:
-            opex -= row["discounted_green_premium"]
-        row["tco"] = opex / INVESTMENT_CYCLE_LENGTH
+            opex -= float(row["discounted_green_premium"])
+        row["tco"] = float(opex / INVESTMENT_CYCLE_LENGTH)
         return row
 
     combined_df["tco"] = 0
